update repo
Maxfashion is an E-Commerce website that was launched in UAE in 2004 while came in India in 2006 and expanded the roots in more than 60 cities of India. The website deals in multiple products for each segment from men to women, child to old; they also provide excellent range of over 3,000 styles every season, in brand new designs, silhouettes and fabrics.

In this project we tried to utilize our learning outcomes of all unit Full Stack Web Development course at Masai School. With our best efforts,I tried to clone the whole website replicating same UI and functionality as of the original with few changes.